pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "llmmap"
version = "1.0.0"
description = "LLMMap - automated LLM prompt injection testing framework"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "PyYAML>=6.0.2",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "pytest-cov>=5.0.0",
  "ruff>=0.6.0",
  "mypy>=1.11.0",
  "types-PyYAML>=6.0.12.20250915",
]

[project.scripts]
llmmap = "llmmap.cli:main"

[tool.setuptools.packages.find]
include = ["llmmap*"]

[tool.setuptools.package-data]
llmmap = ["payloads/packs/*.yaml"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q --cov=llmmap --cov-report=term-missing"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true
strict_optional = true
